third_party/blink/web_tests/platform/linux/virtual/idna-2008/external/wpt/url/toascii.window-expected.txt

This is a testharness.js-based test.
Found 352 tests; 200 PASS, 152 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS Loading data…
PASS aa-- (using URL)
PASS aa-- (using URL.host)
PASS aa-- (using URL.hostname)
PASS aa-- (using <a>)
PASS aa-- (using <a>.host)
PASS aa-- (using <a>.hostname)
PASS aa-- (using <area>)
PASS aa-- (using <area>.host)
PASS aa-- (using <area>.hostname)
PASS a†-- (using URL)
PASS a†-- (using URL.host)
PASS a†-- (using URL.hostname)
PASS a†-- (using <a>)
PASS a†-- (using <a>.host)
PASS a†-- (using <a>.hostname)
PASS a†-- (using <area>)
PASS a†-- (using <area>.host)
PASS a†-- (using <area>.hostname)
PASS ab--c (using URL)
PASS ab--c (using URL.host)
PASS ab--c (using URL.hostname)
PASS ab--c (using <a>)
PASS ab--c (using <a>.host)
PASS ab--c (using <a>.hostname)
PASS ab--c (using <area>)
PASS ab--c (using <area>.host)
PASS ab--c (using <area>.hostname)
PASS -x (using URL)
PASS -x (using URL.host)
PASS -x (using URL.hostname)
PASS -x (using <a>)
PASS -x (using <a>.host)
PASS -x (using <a>.hostname)
PASS -x (using <area>)
PASS -x (using <area>.host)
PASS -x (using <area>.hostname)
PASS -† (using URL)
PASS -† (using URL.host)
PASS -† (using URL.hostname)
PASS -† (using <a>)
PASS -† (using <a>.host)
PASS -† (using <a>.hostname)
PASS -† (using <area>)
PASS -† (using <area>.host)
PASS -† (using <area>.hostname)
PASS -x.xn--zca (using URL)
PASS -x.xn--zca (using URL.host)
PASS -x.xn--zca (using URL.hostname)
PASS -x.xn--zca (using <a>)
PASS -x.xn--zca (using <a>.host)
PASS -x.xn--zca (using <a>.hostname)
PASS -x.xn--zca (using <area>)
PASS -x.xn--zca (using <area>.host)
PASS -x.xn--zca (using <area>.hostname)
FAIL -x.ß (using URL) assert_equals: expected "-x.xn--zca" but got "-x.ss"
FAIL -x.ß (using URL.host) assert_equals: expected "-x.xn--zca" but got "-x.ss"
FAIL -x.ß (using URL.hostname) assert_equals: expected "-x.xn--zca" but got "-x.ss"
FAIL -x.ß (using <a>) assert_equals: expected "-x.xn--zca" but got "-x.ss"
FAIL -x.ß (using <a>.host) assert_equals: expected "-x.xn--zca" but got "-x.ss"
FAIL -x.ß (using <a>.hostname) assert_equals: expected "-x.xn--zca" but got "-x.ss"
FAIL -x.ß (using <area>) assert_equals: expected "-x.xn--zca" but got "-x.ss"
FAIL -x.ß (using <area>.host) assert_equals: expected "-x.xn--zca" but got "-x.ss"
FAIL -x.ß (using <area>.hostname) assert_equals: expected "-x.xn--zca" but got "-x.ss"
PASS x-.xn--zca (using URL)
PASS x-.xn--zca (using URL.host)
PASS x-.xn--zca (using URL.hostname)
PASS x-.xn--zca (using <a>)
PASS x-.xn--zca (using <a>.host)
PASS x-.xn--zca (using <a>.hostname)
PASS x-.xn--zca (using <area>)
PASS x-.xn--zca (using <area>.host)
PASS x-.xn--zca (using <area>.hostname)
FAIL x-.ß (using URL) assert_equals: expected "x-.xn--zca" but got "x-.ss"
FAIL x-.ß (using URL.host) assert_equals: expected "x-.xn--zca" but got "x-.ss"
FAIL x-.ß (using URL.hostname) assert_equals: expected "x-.xn--zca" but got "x-.ss"
FAIL x-.ß (using <a>) assert_equals: expected "x-.xn--zca" but got "x-.ss"
FAIL x-.ß (using <a>.host) assert_equals: expected "x-.xn--zca" but got "x-.ss"
FAIL x-.ß (using <a>.hostname) assert_equals: expected "x-.xn--zca" but got "x-.ss"
FAIL x-.ß (using <area>) assert_equals: expected "x-.xn--zca" but got "x-.ss"
FAIL x-.ß (using <area>.host) assert_equals: expected "x-.xn--zca" but got "x-.ss"
FAIL x-.ß (using <area>.hostname) assert_equals: expected "x-.xn--zca" but got "x-.ss"
PASS x..xn--zca (using URL)
PASS x..xn--zca (using URL.host)
PASS x..xn--zca (using URL.hostname)
PASS x..xn--zca (using <a>)
PASS x..xn--zca (using <a>.host)
PASS x..xn--zca (using <a>.hostname)
PASS x..xn--zca (using <area>)
PASS x..xn--zca (using <area>.host)
PASS x..xn--zca (using <area>.hostname)
FAIL x..ß (using URL) assert_equals: expected "x..xn--zca" but got "x..ss"
FAIL x..ß (using URL.host) assert_equals: expected "x..xn--zca" but got "x..ss"
FAIL x..ß (using URL.hostname) assert_equals: expected "x..xn--zca" but got "x..ss"
FAIL x..ß (using <a>) assert_equals: expected "x..xn--zca" but got "x..ss"
FAIL x..ß (using <a>.host) assert_equals: expected "x..xn--zca" but got "x..ss"
FAIL x..ß (using <a>.hostname) assert_equals: expected "x..xn--zca" but got "x..ss"
FAIL x..ß (using <area>) assert_equals: expected "x..xn--zca" but got "x..ss"
FAIL x..ß (using <area>.host) assert_equals: expected "x..xn--zca" but got "x..ss"
FAIL x..ß (using <area>.hostname) assert_equals: expected "x..xn--zca" but got "x..ss"
FAIL xn--a (using URL) assert_throws_js: function "() => makeURL("url", hostTest.input)" did not throw
FAIL xn--a (using URL.host) assert_equals: expected "x" but got "xn--a"
FAIL xn--a (using URL.hostname) assert_equals: expected "x" but got "xn--a"
FAIL xn--a (using <a>) assert_equals: expected "" but got "xn--a"
FAIL xn--a (using <a>.host) assert_equals: expected "x" but got "xn--a"
FAIL xn--a (using <a>.hostname) assert_equals: expected "x" but got "xn--a"
FAIL xn--a (using <area>) assert_equals: expected "" but got "xn--a"
FAIL xn--a (using <area>.host) assert_equals: expected "x" but got "xn--a"
FAIL xn--a (using <area>.hostname) assert_equals: expected "x" but got "xn--a"
FAIL xn--a.xn--zca (using URL) assert_throws_js: function "() => makeURL("url", hostTest.input)" did not throw
FAIL xn--a.xn--zca (using URL.host) assert_equals: expected "x" but got "xn--a.xn--zca"
FAIL xn--a.xn--zca (using URL.hostname) assert_equals: expected "x" but got "xn--a.xn--zca"
FAIL xn--a.xn--zca (using <a>) assert_equals: expected "" but got "xn--a.xn--zca"
FAIL xn--a.xn--zca (using <a>.host) assert_equals: expected "x" but got "xn--a.xn--zca"
FAIL xn--a.xn--zca (using <a>.hostname) assert_equals: expected "x" but got "xn--a.xn--zca"
FAIL xn--a.xn--zca (using <area>) assert_equals: expected "" but got "xn--a.xn--zca"
FAIL xn--a.xn--zca (using <area>.host) assert_equals: expected "x" but got "xn--a.xn--zca"
FAIL xn--a.xn--zca (using <area>.hostname) assert_equals: expected "x" but got "xn--a.xn--zca"
PASS xn--a.ß (using URL)
PASS xn--a.ß (using URL.host)
PASS xn--a.ß (using URL.hostname)
FAIL xn--a.ß (using <a>) assert_equals: expected "https://xn--a.ß/x" but got "https://xn--a.%C3%9F/x"
PASS xn--a.ß (using <a>.host)
PASS xn--a.ß (using <a>.hostname)
FAIL xn--a.ß (using <area>) assert_equals: expected "https://xn--a.ß/x" but got "https://xn--a.%C3%9F/x"
PASS xn--a.ß (using <area>.host)
PASS xn--a.ß (using <area>.hostname)
FAIL xn--ls8h= (using URL) assert_throws_js: function "() => makeURL("url", hostTest.input)" did not throw
FAIL xn--ls8h= (using URL.host) assert_equals: expected "x" but got "xn--ls8h%3D"
FAIL xn--ls8h= (using URL.hostname) assert_equals: expected "x" but got "xn--ls8h%3D"
FAIL xn--ls8h= (using <a>) assert_equals: expected "" but got "xn--ls8h%3D"
FAIL xn--ls8h= (using <a>.host) assert_equals: expected "x" but got "xn--ls8h%3D"
FAIL xn--ls8h= (using <a>.hostname) assert_equals: expected "x" but got "xn--ls8h%3D"
FAIL xn--ls8h= (using <area>) assert_equals: expected "" but got "xn--ls8h%3D"
FAIL xn--ls8h= (using <area>.host) assert_equals: expected "x" but got "xn--ls8h%3D"
FAIL xn--ls8h= (using <area>.hostname) assert_equals: expected "x" but got "xn--ls8h%3D"
PASS xn--tešla (using URL)
PASS xn--tešla (using URL.host)
PASS xn--tešla (using URL.hostname)
FAIL xn--tešla (using <a>) assert_equals: expected "https://xn--tešla/x" but got "https://xn--te%C5%A1la/x"
PASS xn--tešla (using <a>.host)
PASS xn--tešla (using <a>.hostname)
FAIL xn--tešla (using <area>) assert_equals: expected "https://xn--tešla/x" but got "https://xn--te%C5%A1la/x"
PASS xn--tešla (using <area>.host)
PASS xn--tešla (using <area>.hostname)
PASS xn--zca.xn--zca (using URL)
PASS xn--zca.xn--zca (using URL.host)
PASS xn--zca.xn--zca (using URL.hostname)
PASS xn--zca.xn--zca (using <a>)
PASS xn--zca.xn--zca (using <a>.host)
PASS xn--zca.xn--zca (using <a>.hostname)
PASS xn--zca.xn--zca (using <area>)
PASS xn--zca.xn--zca (using <area>.host)
PASS xn--zca.xn--zca (using <area>.hostname)
FAIL xn--zca.ß (using URL) assert_equals: expected "xn--zca.xn--zca" but got "xn--zca.ss"
FAIL xn--zca.ß (using URL.host) assert_equals: expected "xn--zca.xn--zca" but got "xn--zca.ss"
FAIL xn--zca.ß (using URL.hostname) assert_equals: expected "xn--zca.xn--zca" but got "xn--zca.ss"
FAIL xn--zca.ß (using <a>) assert_equals: expected "xn--zca.xn--zca" but got "xn--zca.ss"
FAIL xn--zca.ß (using <a>.host) assert_equals: expected "xn--zca.xn--zca" but got "xn--zca.ss"
FAIL xn--zca.ß (using <a>.hostname) assert_equals: expected "xn--zca.xn--zca" but got "xn--zca.ss"
FAIL xn--zca.ß (using <area>) assert_equals: expected "xn--zca.xn--zca" but got "xn--zca.ss"
FAIL xn--zca.ß (using <area>.host) assert_equals: expected "xn--zca.xn--zca" but got "xn--zca.ss"
FAIL xn--zca.ß (using <area>.hostname) assert_equals: expected "xn--zca.xn--zca" but got "xn--zca.ss"
PASS ab--c.xn--zca (using URL)
PASS ab--c.xn--zca (using URL.host)
PASS ab--c.xn--zca (using URL.hostname)
PASS ab--c.xn--zca (using <a>)
PASS ab--c.xn--zca (using <a>.host)
PASS ab--c.xn--zca (using <a>.hostname)
PASS ab--c.xn--zca (using <area>)
PASS ab--c.xn--zca (using <area>.host)
PASS ab--c.xn--zca (using <area>.hostname)
FAIL ab--c.ß (using URL) assert_equals: expected "ab--c.xn--zca" but got "ab--c.ss"
FAIL ab--c.ß (using URL.host) assert_equals: expected "ab--c.xn--zca" but got "ab--c.ss"
FAIL ab--c.ß (using URL.hostname) assert_equals: expected "ab--c.xn--zca" but got "ab--c.ss"
FAIL ab--c.ß (using <a>) assert_equals: expected "ab--c.xn--zca" but got "ab--c.ss"
FAIL ab--c.ß (using <a>.host) assert_equals: expected "ab--c.xn--zca" but got "ab--c.ss"
FAIL ab--c.ß (using <a>.hostname) assert_equals: expected "ab--c.xn--zca" but got "ab--c.ss"
FAIL ab--c.ß (using <area>) assert_equals: expected "ab--c.xn--zca" but got "ab--c.ss"
FAIL ab--c.ß (using <area>.host) assert_equals: expected "ab--c.xn--zca" but got "ab--c.ss"
FAIL ab--c.ß (using <area>.hostname) assert_equals: expected "ab--c.xn--zca" but got "ab--c.ss"
FAIL ‍.example (using URL) assert_throws_js: function "() => makeURL("url", hostTest.input)" did not throw
FAIL ‍.example (using URL.host) assert_equals: expected "x" but got ".example"
FAIL ‍.example (using URL.hostname) assert_equals: expected "x" but got ".example"
FAIL ‍.example (using <a>) assert_equals: expected "" but got ".example"
FAIL ‍.example (using <a>.host) assert_equals: expected "x" but got ".example"
FAIL ‍.example (using <a>.hostname) assert_equals: expected "x" but got ".example"
FAIL ‍.example (using <area>) assert_equals: expected "" but got ".example"
FAIL ‍.example (using <area>.host) assert_equals: expected "x" but got ".example"
FAIL ‍.example (using <area>.hostname) assert_equals: expected "x" but got ".example"
FAIL xn--1ug.example (using URL) assert_throws_js: function "() => makeURL("url", hostTest.input)" did not throw
FAIL xn--1ug.example (using URL.host) assert_equals: expected "x" but got "xn--1ug.example"
FAIL xn--1ug.example (using URL.hostname) assert_equals: expected "x" but got "xn--1ug.example"
FAIL xn--1ug.example (using <a>) assert_equals: expected "" but got "xn--1ug.example"
FAIL xn--1ug.example (using <a>.host) assert_equals: expected "x" but got "xn--1ug.example"
FAIL xn--1ug.example (using <a>.hostname) assert_equals: expected "x" but got "xn--1ug.example"
FAIL xn--1ug.example (using <area>) assert_equals: expected "" but got "xn--1ug.example"
FAIL xn--1ug.example (using <area>.host) assert_equals: expected "x" but got "xn--1ug.example"
FAIL xn--1ug.example (using <area>.hostname) assert_equals: expected "x" but got "xn--1ug.example"
PASS يa (using URL)
PASS يa (using URL.host)
PASS يa (using URL.hostname)
FAIL يa (using <a>) assert_equals: expected "https://يa/x" but got "https://%D9%8Aa/x"
PASS يa (using <a>.host)
PASS يa (using <a>.hostname)
FAIL يa (using <area>) assert_equals: expected "https://يa/x" but got "https://%D9%8Aa/x"
PASS يa (using <area>.host)
PASS يa (using <area>.hostname)
FAIL xn--a-yoc (using URL) assert_throws_js: function "() => makeURL("url", hostTest.input)" did not throw
FAIL xn--a-yoc (using URL.host) assert_equals: expected "x" but got "xn--a-yoc"
FAIL xn--a-yoc (using URL.hostname) assert_equals: expected "x" but got "xn--a-yoc"
FAIL xn--a-yoc (using <a>) assert_equals: expected "" but got "xn--a-yoc"
FAIL xn--a-yoc (using <a>.host) assert_equals: expected "x" but got "xn--a-yoc"
FAIL xn--a-yoc (using <a>.hostname) assert_equals: expected "x" but got "xn--a-yoc"
FAIL xn--a-yoc (using <area>) assert_equals: expected "" but got "xn--a-yoc"
FAIL xn--a-yoc (using <area>.host) assert_equals: expected "x" but got "xn--a-yoc"
FAIL xn--a-yoc (using <area>.hostname) assert_equals: expected "x" but got "xn--a-yoc"
FAIL ශ්‍රී (using URL) assert_equals: expected "xn--10cl1a0b660p" but got "xn--10cl1a0b"
FAIL ශ්‍රී (using URL.host) assert_equals: expected "xn--10cl1a0b660p" but got "xn--10cl1a0b"
FAIL ශ්‍රී (using URL.hostname) assert_equals: expected "xn--10cl1a0b660p" but got "xn--10cl1a0b"
FAIL ශ්‍රී (using <a>) assert_equals: expected "xn--10cl1a0b660p" but got "xn--10cl1a0b"
FAIL ශ්‍රී (using <a>.host) assert_equals: expected "xn--10cl1a0b660p" but got "xn--10cl1a0b"
FAIL ශ්‍රී (using <a>.hostname) assert_equals: expected "xn--10cl1a0b660p" but got "xn--10cl1a0b"
FAIL ශ්‍රී (using <area>) assert_equals: expected "xn--10cl1a0b660p" but got "xn--10cl1a0b"
FAIL ශ්‍රී (using <area>.host) assert_equals: expected "xn--10cl1a0b660p" but got "xn--10cl1a0b"
FAIL ශ්‍රී (using <area>.hostname) assert_equals: expected "xn--10cl1a0b660p" but got "xn--10cl1a0b"
FAIL نامه‌ای (using URL) assert_equals: expected "xn--mgba3gch31f060k" but got "xn--mgba3gch31f"
FAIL نامه‌ای (using URL.host) assert_equals: expected "xn--mgba3gch31f060k" but got "xn--mgba3gch31f"
FAIL نامه‌ای (using URL.hostname) assert_equals: expected "xn--mgba3gch31f060k" but got "xn--mgba3gch31f"
FAIL نامه‌ای (using <a>) assert_equals: expected "xn--mgba3gch31f060k" but got "xn--mgba3gch31f"
FAIL نامه‌ای (using <a>.host) assert_equals: expected "xn--mgba3gch31f060k" but got "xn--mgba3gch31f"
FAIL نامه‌ای (using <a>.hostname) assert_equals: expected "xn--mgba3gch31f060k" but got "xn--mgba3gch31f"
FAIL نامه‌ای (using <area>) assert_equals: expected "xn--mgba3gch31f060k" but got "xn--mgba3gch31f"
FAIL نامه‌ای (using <area>.host) assert_equals: expected "xn--mgba3gch31f060k" but got "xn--mgba3gch31f"
FAIL نامه‌ای (using <area>.hostname) assert_equals: expected "xn--mgba3gch31f060k" but got "xn--mgba3gch31f"
PASS �.com (using URL)
PASS �.com (using URL.host)
PASS �.com (using URL.hostname)
FAIL �.com (using <a>) assert_equals: expected "https://\ufffd.com/x" but got "https://%EF%BF%BD.com/x"
PASS �.com (using <a>.host)
PASS �.com (using <a>.hostname)
FAIL �.com (using <area>) assert_equals: expected "https://\ufffd.com/x" but got "https://%EF%BF%BD.com/x"
PASS �.com (using <area>.host)
PASS �.com (using <area>.hostname)
FAIL xn--zn7c.com (using URL) assert_throws_js: function "() => makeURL("url", hostTest.input)" did not throw
FAIL xn--zn7c.com (using URL.host) assert_equals: expected "x" but got "xn--zn7c.com"
FAIL xn--zn7c.com (using URL.hostname) assert_equals: expected "x" but got "xn--zn7c.com"
FAIL xn--zn7c.com (using <a>) assert_equals: expected "" but got "xn--zn7c.com"
FAIL xn--zn7c.com (using <a>.host) assert_equals: expected "x" but got "xn--zn7c.com"
FAIL xn--zn7c.com (using <a>.hostname) assert_equals: expected "x" but got "xn--zn7c.com"
FAIL xn--zn7c.com (using <area>) assert_equals: expected "" but got "xn--zn7c.com"
FAIL xn--zn7c.com (using <area>.host) assert_equals: expected "x" but got "xn--zn7c.com"
FAIL xn--zn7c.com (using <area>.hostname) assert_equals: expected "x" but got "xn--zn7c.com"
PASS x01234567890123456789012345678901234567890123456789012345678901x (using URL)
PASS x01234567890123456789012345678901234567890123456789012345678901x (using URL.host)
PASS x01234567890123456789012345678901234567890123456789012345678901x (using URL.hostname)
PASS x01234567890123456789012345678901234567890123456789012345678901x (using <a>)
PASS x01234567890123456789012345678901234567890123456789012345678901x (using <a>.host)
PASS x01234567890123456789012345678901234567890123456789012345678901x (using <a>.hostname)
PASS x01234567890123456789012345678901234567890123456789012345678901x (using <area>)
PASS x01234567890123456789012345678901234567890123456789012345678901x (using <area>.host)
PASS x01234567890123456789012345678901234567890123456789012345678901x (using <area>.hostname)
PASS x01234567890123456789012345678901234567890123456789012345678901† (using URL)
PASS x01234567890123456789012345678901234567890123456789012345678901† (using URL.host)
PASS x01234567890123456789012345678901234567890123456789012345678901† (using URL.hostname)
PASS x01234567890123456789012345678901234567890123456789012345678901† (using <a>)
PASS x01234567890123456789012345678901234567890123456789012345678901† (using <a>.host)
PASS x01234567890123456789012345678901234567890123456789012345678901† (using <a>.hostname)
PASS x01234567890123456789012345678901234567890123456789012345678901† (using <area>)
PASS x01234567890123456789012345678901234567890123456789012345678901† (using <area>.host)
PASS x01234567890123456789012345678901234567890123456789012345678901† (using <area>.hostname)
PASS x01234567890123456789012345678901234567890123456789012345678901x.xn--zca (using URL)
PASS x01234567890123456789012345678901234567890123456789012345678901x.xn--zca (using URL.host)
PASS x01234567890123456789012345678901234567890123456789012345678901x.xn--zca (using URL.hostname)
PASS x01234567890123456789012345678901234567890123456789012345678901x.xn--zca (using <a>)
PASS x01234567890123456789012345678901234567890123456789012345678901x.xn--zca (using <a>.host)
PASS x01234567890123456789012345678901234567890123456789012345678901x.xn--zca (using <a>.hostname)
PASS x01234567890123456789012345678901234567890123456789012345678901x.xn--zca (using <area>)
PASS x01234567890123456789012345678901234567890123456789012345678901x.xn--zca (using <area>.host)
PASS x01234567890123456789012345678901234567890123456789012345678901x.xn--zca (using <area>.hostname)
FAIL x01234567890123456789012345678901234567890123456789012345678901x.ß (using URL) assert_equals: expected "x01234567890123456789012345678901234567890123456789012345678901x.xn--zca" but got "x01234567890123456789012345678901234567890123456789012345678901x.ss"
FAIL x01234567890123456789012345678901234567890123456789012345678901x.ß (using URL.host) assert_equals: expected "x01234567890123456789012345678901234567890123456789012345678901x.xn--zca" but got "x01234567890123456789012345678901234567890123456789012345678901x.ss"
FAIL x01234567890123456789012345678901234567890123456789012345678901x.ß (using URL.hostname) assert_equals: expected "x01234567890123456789012345678901234567890123456789012345678901x.xn--zca" but got "x01234567890123456789012345678901234567890123456789012345678901x.ss"
FAIL x01234567890123456789012345678901234567890123456789012345678901x.ß (using <a>) assert_equals: expected "x01234567890123456789012345678901234567890123456789012345678901x.xn--zca" but got "x01234567890123456789012345678901234567890123456789012345678901x.ss"
FAIL x01234567890123456789012345678901234567890123456789012345678901x.ß (using <a>.host) assert_equals: expected "x01234567890123456789012345678901234567890123456789012345678901x.xn--zca" but got "x01234567890123456789012345678901234567890123456789012345678901x.ss"
FAIL x01234567890123456789012345678901234567890123456789012345678901x.ß (using <a>.hostname) assert_equals: expected "x01234567890123456789012345678901234567890123456789012345678901x.xn--zca" but got "x01234567890123456789012345678901234567890123456789012345678901x.ss"
FAIL x01234567890123456789012345678901234567890123456789012345678901x.ß (using <area>) assert_equals: expected "x01234567890123456789012345678901234567890123456789012345678901x.xn--zca" but got "x01234567890123456789012345678901234567890123456789012345678901x.ss"
FAIL x01234567890123456789012345678901234567890123456789012345678901x.ß (using <area>.host) assert_equals: expected "x01234567890123456789012345678901234567890123456789012345678901x.xn--zca" but got "x01234567890123456789012345678901234567890123456789012345678901x.ss"
FAIL x01234567890123456789012345678901234567890123456789012345678901x.ß (using <area>.hostname) assert_equals: expected "x01234567890123456789012345678901234567890123456789012345678901x.xn--zca" but got "x01234567890123456789012345678901234567890123456789012345678901x.ss"
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.x (using URL)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.x (using URL.host)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.x (using URL.hostname)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.x (using <a>)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.x (using <a>.host)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.x (using <a>.hostname)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.x (using <area>)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.x (using <area>.host)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.x (using <area>.hostname)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--zca (using URL)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--zca (using URL.host)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--zca (using URL.hostname)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--zca (using <a>)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--zca (using <a>.host)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--zca (using <a>.hostname)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--zca (using <area>)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--zca (using <area>.host)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--zca (using <area>.hostname)
FAIL 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.ß (using URL) assert_equals: expected "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--zca" but got "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.ss"
FAIL 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.ß (using URL.host) assert_equals: expected "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--zca" but got "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.ss"
FAIL 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.ß (using URL.hostname) assert_equals: expected "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--zca" but got "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.ss"
FAIL 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.ß (using <a>) assert_equals: expected "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--zca" but got "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.ss"
FAIL 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.ß (using <a>.host) assert_equals: expected "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--zca" but got "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.ss"
FAIL 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.ß (using <a>.hostname) assert_equals: expected "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--zca" but got "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.ss"
FAIL 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.ß (using <area>) assert_equals: expected "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--zca" but got "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.ss"
FAIL 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.ß (using <area>.host) assert_equals: expected "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--zca" but got "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.ss"
FAIL 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.ß (using <area>.hostname) assert_equals: expected "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--zca" but got "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.ss"
PASS a­b (using URL)
PASS a­b (using URL.host)
PASS a­b (using URL.hostname)
PASS a­b (using <a>)
PASS a­b (using <a>.host)
PASS a­b (using <a>.hostname)
PASS a­b (using <area>)
PASS a­b (using <area>.host)
PASS a­b (using <area>.hostname)
PASS ≠ (using URL)
PASS ≠ (using URL.host)
PASS ≠ (using URL.hostname)
PASS ≠ (using <a>)
PASS ≠ (using <a>.host)
PASS ≠ (using <a>.hostname)
PASS ≠ (using <area>)
PASS ≠ (using <area>.host)
PASS ≠ (using <area>.hostname)
PASS ≮ (using URL)
PASS ≮ (using URL.host)
PASS ≮ (using URL.hostname)
PASS ≮ (using <a>)
PASS ≮ (using <a>.host)
PASS ≮ (using <a>.hostname)
PASS ≮ (using <area>)
PASS ≮ (using <area>.host)
PASS ≮ (using <area>.hostname)
PASS ≯ (using URL)
PASS ≯ (using URL.host)
PASS ≯ (using URL.hostname)
PASS ≯ (using <a>)
PASS ≯ (using <a>.host)
PASS ≯ (using <a>.hostname)
PASS ≯ (using <area>)
PASS ≯ (using <area>.host)
PASS ≯ (using <area>.hostname)
Harness: the test ran to completion.
